better centering for crosshair

+++ src/gui/image_display.c
@@ -604,8 +604,8 @@ static void draw_selection(const draw_data_t* dd) {
 		// display a mini cross when the selection is being dragged
 		if (com.freezeX && com.freezeY) {
 			cairo_set_line_width(cr, 1.0 / dd->zoom);
-			point selection_center = { com.selection.x + com.selection.w / 2,
-				com.selection.y + com.selection.h / 2 };
+			point selection_center = { com.selection.x + (double)com.selection.w / 2.,
+				com.selection.y + (double)com.selection.h / 2. };
 			cairo_move_to(cr, selection_center.x, selection_center.y - 5 / dd->zoom);
 			cairo_line_to(cr, selection_center.x, selection_center.y + 5 / dd->zoom);
 			cairo_move_to(cr, selection_center.x - 5 / dd->zoom, selection_center.y);
